mythpreviewgen: rename commandlineparser.* to mythpreviewgen_…
