+++ mythtv/programs/mythpreviewgen/mythpreviewgen_commandlineparser.cpp
@@ -1,7 +1,7 @@
 #include <QString>
 
 #include "mythcorecontext.h"
-#include "commandlineparser.h"
+#include "mythpreviewgen_commandlineparser.h"
 
 MythPreviewGeneratorCommandLineParser::MythPreviewGeneratorCommandLineParser() :
     MythCommandLineParser(MYTH_APPNAME_MYTHPREVIEWGEN)
